feat: complete day 11, puzzle 2

+++ src/AdventOfCode2023.Day11/Usings.cs
@@ -1,1 +1,1 @@
-global using Point = (int x, int y);
+global using Point = (long x, long y);
